start by pulling stuff out of app.js
